libcxx: Prevent cmake from using system llvm-config
Pass -DLLVM_CONFIG=NO to the CMake process for libcxx in order to
prevent it from finding the system-wide install of llvm-config. This
provides a consistent behavior across all hosts, independently of
whether LLVM is installed, and also fixes build failure on Gentoo where
using 'llvm-config --src-root' on installed instance is prohibited.